+++ lib/rspec/activerecord/expectations.rb
@@ -9,5 +9,6 @@ module RSpec
 end
 
 require_relative 'expectations/errors'
+require_relative 'expectations/query_inspector'
 require_relative 'expectations/collector'
 require_relative 'expectations/matchers/query_count'