+++ lib/rspec/activerecord/expectations/collector.rb
@@ -1,15 +1,16 @@
 module RSpec::ActiveRecord::Expectations
   class Collector
     def initialize
-      @counts = {
-        queries: 0
-      }
+      @inspector  = QueryInspector.new
+      @counts     = QueryInspector.valid_query_types.each_with_object({}) do |query_type, hash|
+        hash[query_type] = 0
+      end
 
       ActiveSupport::Notifications.subscribe("sql.active_record", method(:record_query))
     end
 
     def queries_of_type(type)
-      @counts.fetch(type)
+      @counts[type] || (raise ArgumentError, "Sorry, #{type} is not a valid kind of query")
     end
 
     def valid_type?(type)
@@ -17,7 +18,10 @@ module RSpec::ActiveRecord::Expectations
     end
 
     def record_query(*_unused, data)
-      @counts[:queries] += 1
+      categories = @inspector.categorize(data)
+      categories.each do |category|
+        @counts[category] += 1
+      end
     end
   end
 end